+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 * Migrate to null safety.
 * Fix outdated URLs in `README.md`.
 * BREAKING: Removed archive support.
+* BREAKING: `DirectoryDescriptor.load` only supports a `String` path instead of
+  also accepting relative `Uri` objects.
 
 ## 1.2.0